break up compilation pass
