+++ src/ir_passes/mod.rs
@@ -1,15 +1,15 @@
 mod assume;
-mod compile;
 mod discharge;
 mod hoist_facts;
 mod interval_check;
+mod lower;
 mod prop_simplify;
 mod type_check;
 
 pub use assume::Assume;
-pub use compile::Compile;
 pub use discharge::Discharge;
 pub use hoist_facts::HoistFacts;
 pub use interval_check::IntervalCheck;
+pub use lower::Compile;
 pub use prop_simplify::Simplify;
 pub use type_check::TypeCheck;
